requirements.txt
smartapi-python
pycryptodome
fastapi
uvicorn
gunicorn
websocket-client
pyotp
logzero
